FetchWeatherErrorの切り出し
